pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "morningstar"
version = "0.1.0"
description = "A sardonic deliberative coding partner."
readme = "README.md"
authors = [
  { name = "Lucius J. Morningstar", email = "lucius@example.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "click>=8.0.0",
    "jsonschema>=4.0.0",
    "python-dateutil>=2.8.2",
]

[project.scripts]
morningstar = "tools.cli:cli"
